Move _AdapterTester to openscm_runner.testing
