+++ tests/integration/test_fair1X.py
@@ -3,11 +3,11 @@ import os.path
 import numpy as np
 import numpy.testing as npt
 import pytest
-from base import _AdapterTester
 from scmdata import ScmRun
 
 from openscm_runner import run
 from openscm_runner.adapters import FAIR
+from openscm_runner.testing import _AdapterTester
 from openscm_runner.utils import calculate_quantiles
