+++ tests/integration/test_magicc7.py
@@ -2,11 +2,11 @@ import os.path
 
 import pymagicc.io
 import pytest
-from base import _AdapterTester
 from scmdata import ScmRun
 
 from openscm_runner import run
 from openscm_runner.adapters import MAGICC7
+from openscm_runner.testing import _AdapterTester
 from openscm_runner.utils import calculate_quantiles
